[refactor] service 인터페이스 삭제 (구현체가 하나 이상일 경우가 없으므로)
